import sycamore does not import ray
import ray is about 0.2s, import sycamore about 3s, so ~10% speedup.
