+++ lib/sycamore/sycamore/connectors/file/file_writer.py
@@ -1,25 +1,19 @@
-from sycamore.data import Document, MetadataDocument
+from sycamore.data import Document
 from sycamore.plan_nodes import Node, Write
 
-from pyarrow.fs import FileSystem, FileType
-from pyarrow import NativeFile
-
-from ray.data import Dataset
-from ray.data.datasource import FilenameProvider, Datasink, BlockBasedFileDatasink
-from ray.data.datasource.path_util import _resolve_paths_and_filesystem
-from ray.data.block import Block, BlockAccessor
-from ray.data._internal.execution.interfaces import TaskContext
+from pyarrow.fs import FileSystem
 
 from collections import UserDict
 from io import StringIO
 import json
 import logging
 from pathlib import Path
-import posixpath
-from urllib.parse import urlparse
 import uuid
-from typing import Any, Callable, Optional, Iterable
-from sycamore.utils.time_trace import TimeTrace
+from typing import Callable, Optional, TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from ray.data import Dataset
+
 
 logger = logging.getLogger(__name__)
 
@@ -155,7 +149,9 @@ class FileWriter(Write):
         self.doc_to_bytes_fn = doc_to_bytes_fn
         self.ray_remote_args = ray_remote_args
 
-    def execute(self, **kwargs) -> Dataset:
+    def execute(self, **kwargs) -> "Dataset":
+        from sycamore.connectors.file.file_writer_ray import _FileDataSink
+
         dataset = self.child().execute()
 
         dataset.write_datasink(
@@ -201,80 +197,10 @@ class JsonWriter(Write):
         self.filesystem = filesystem
         self.ray_remote_args = ray_remote_args
 
-    def execute(self, **kwargs) -> Dataset:
+    def execute(self, **kwargs) -> "Dataset":
         ds = self.child().execute()
+        from sycamore.connectors.file.file_writer_ray import _JsonBlockDataSink
+
         sink = _JsonBlockDataSink(self.path, filesystem=self.filesystem)
         ds.write_datasink(sink, ray_remote_args=self.ray_remote_args)
         return ds
-
-
-class DocToRowFilenameProvider(FilenameProvider):
-    def __init__(self, filename_fn: Callable[[Document], str]):
-        self._filename_fn = filename_fn
-
-    def get_filename_for_row(self, row: dict[str, Any], task_index: int, block_index: int, row_index: int) -> str:
-        return self._filename_fn(Document.from_row(row))
-
-
-class BlockFilenameProvider(FilenameProvider):
-    def get_filename_for_block(self, block: Block, task_index: int, block_index: int) -> str:
-        return f"block_{block_index}_{task_index}.jsonl"
-
-
-class _FileDataSink(Datasink):
-    def __init__(
-        self,
-        path: str,
-        filesystem: Optional[FileSystem] = None,
-        filename_fn: Callable[[Document], str] = default_filename,
-        doc_to_bytes_fn: Callable[[Document], bytes] = default_doc_to_bytes,
-        makedirs: bool = True,
-    ):
-        (paths, self._filesystem) = _resolve_paths_and_filesystem(path, filesystem)
-        self._root = paths[0]
-        self._filename_fn = filename_fn
-        self._doc_to_bytes_fn = doc_to_bytes_fn
-        self._makedirs = makedirs
-
-    def on_write_start(self) -> None:
-        if not self._makedirs:
-            return
-
-        # This follows Ray logic to skip attempting to
-        # create "directories" for s3 filesystems.
-        parsed_uri = urlparse(self._root)
-        is_s3_uri = parsed_uri.scheme == "s3"
-
-        if not is_s3_uri and self._filesystem.get_file_info(self._root).type is FileType.NotFound:
-            self._filesystem.create_dir(self._root, recursive=True)
-
-    def write(self, blocks: Iterable[Block], ctx: TaskContext) -> Any:
-        for block in blocks:
-            b = BlockAccessor.for_block(block).to_arrow().to_pylist()
-            for _, row in enumerate(b):
-                doc = Document.from_row(row)
-                if isinstance(doc, MetadataDocument):
-                    continue
-                bytes = self._doc_to_bytes_fn(doc)
-                path = posixpath.join(self._root, self._filename_fn(doc))
-                with self._filesystem.open_output_stream(path) as file:
-                    file.write(bytes)
-
-
-class _JsonBlockDataSink(BlockBasedFileDatasink):
-    def __init__(
-        self,
-        path: str,
-        filesystem: Optional[FileSystem] = None,
-    ) -> None:
-        super().__init__(path, filesystem=filesystem, filename_provider=BlockFilenameProvider())
-
-    def write_block_to_file(self, block: BlockAccessor, file: NativeFile) -> None:
-        with TimeTrace("jsonSink"):
-            for row in block.iter_rows(True):  # type: ignore[var-annotated]
-                doc = Document.from_row(row)
-                if isinstance(doc, MetadataDocument):
-                    continue
-                del doc.binary_representation  # Doesn't make sense in JSON
-                binary = document_to_bytes(doc)
-                file.write(binary)
