+++ lib/sycamore/sycamore/transforms/partition.py
@@ -4,8 +4,6 @@ from typing import Any, Optional
 
 from bs4 import BeautifulSoup
 
-from ray.data import ActorPoolStrategy
-
 from sycamore.functions import TextOverlapChunker, Chunker
 from sycamore.functions import CharacterTokenizer, Tokenizer
 from sycamore.functions import reorder_elements
@@ -606,6 +604,8 @@ class Partition(CompositeTransform):
         self, child: Node, partitioner: Partitioner, table_extractor: Optional[TableExtractor] = None, **resource_args
     ):
         ops = []
+        from ray.data import ActorPoolStrategy
+
         if isinstance(partitioner, ArynPartitioner) and partitioner._use_partitioning_service:
             resource_args["compute"] = ActorPoolStrategy(size=1)
         if partitioner.device == "cuda":
