+++ lib/sycamore/sycamore/transforms/sketcher.py
@@ -3,8 +3,6 @@ import re
 import functools
 import unicodedata
 
-from ray.data import ActorPoolStrategy
-
 from sycamore.data import Document
 from sycamore.functions.simhash import shinglesCalc, shinglesDist
 from sycamore.plan_nodes import Node, SingleThreadUser, NonGPUUser
@@ -84,6 +82,8 @@ class SketchUniquify(SingleThreadUser, NonGPUUser, FlatMap):
 
     def __init__(self, child: Node, threshold: float = 0.4, **kwargs) -> None:
         # must run on 1 instance to get a global view
+        from ray.data import ActorPoolStrategy
+
         kwargs["compute"] = ActorPoolStrategy(size=1)
         super().__init__(child, f=SketchUniquify.Predicate, constructor_args=[threshold], **kwargs)
 
@@ -137,6 +137,8 @@ class SketchDebug(SingleThreadUser, NonGPUUser, FlatMap):
     """
 
     def __init__(self, child: Node, threshold: float = 0.4, **kwargs) -> None:
+        from ray.data import ActorPoolStrategy
+
         kwargs["compute"] = ActorPoolStrategy(size=1)
         super().__init__(child, f=SketchDebug.Predicate, constructor_args=[threshold], **kwargs)
         self.threshold = threshold
